Arbitrage Sports/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arbitrage-sports"
version = "0.1.0"
description = "Alert-only sportsbook vs Kalshi edge scanner"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27,<0.29",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
